模式引入值域id获取

+++ kgms/src/main/java/com/plantdata/kgcloud/domain/dw/service/impl/PreBuilderServiceImpl.java
@@ -1489,7 +1489,7 @@ public class PreBuilderServiceImpl implements PreBuilderService {
                 List<Long> rangeIds = new ArrayList<>();
                 if(modelRanges != null && !modelRanges.isEmpty()){
                     for(ModelRangeRsp rangeRsp: modelRanges){
-                        rangeIds.add(conceptIdMap.get(rangeRsp.getRange()));
+                        rangeIds.add(conceptIdMap.get(rangeRsp.getRange().intValue()));
                     }
                 }
